pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-theme-slate"
description = "Bootstrap Theme 'Slate' for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "eveonline",
    "template",
    "theme",
]
license = { file = "LICENSE" }
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.10"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.10,<5",
    "allianceauth-app-utils>=1.25",
]
optional-dependencies.tests-allianceauth-latest = [
    "coverage",
    "django-webtest",
]
urls.Changelog = "https://github.com/ppfeufer/aa-theme-slate/blob/master/CHANGELOG.md"
urls.Documentation = "https://github.com/ppfeufer/aa-theme-slate/blob/master/README.md"
urls.Donations = "https://ko-fi.com/ppfeufer"
urls.Homepage = "https://github.com/ppfeufer/aa-theme-slate"
urls.Source = "https://github.com/ppfeufer/aa-theme-slate.git"
urls.Tracker = "https://github.com/ppfeufer/aa-theme-slate/issues"

[tool.hatch.build]
include = [
    "/aa_theme_slate",
]

[tool.hatch.version]
path = "aa_theme_slate/__init__.py"
